requirements.txt
aiofiles==23.2.1
altair==5.2.0
anyio==4.3.0
aspose-words==24.3.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==6.0.1
decorator==4.4.2
etelemetry==0.3.1
filelock==3.13.1
fitz==0.0.1.dev2
frontend==0.0.3
gitdb==4.0.11
GitPython==3.1.42
h11==0.14.0
httplib2==0.22.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
looseversion==1.3.0
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
moviepy==1.0.3
mysql-connector-python==8.3.0
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
pathlib==1.0.1
pillow==10.2.0
proglog==0.1.10
protobuf==4.25.3
prov==2.0.0
pyarrow==15.0.2
pydeck==0.8.1b0
pydot==2.0.0
pydub==0.25.1
Pygments==2.17.2
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
pyparsing==3.1.2
PyPDF2==3.0.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pyttsx3==2.90
pytz==2024.1
pyxnat==1.6.2
rdflib==7.0.0
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scipy==1.12.0
selenium==4.18.1
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
SpeechRecognition==3.10.1
starlette==0.37.2
streamlit==1.32.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traits==6.3.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
watchdog==4.0.0
wsproto==1.2.0
